Home Screen & Mentee Screen Update
Home screen now has a dynamic list using "menteelist.js" which tentatively uses a mentee array.
For each mentee in the array, it creates a pressable area in which it grabs the mentee's values of name and risk in order to display the name and correct butterfly icon. List updates and changes based on what is in the array.
each mentee's pressable goes to "menteescreen.js", passing itself (the mentee) in order for menteescreen.js to display that mentees information
 "<MenteeList navigation={navigation)>" is then called in "homescreen.js".

When a mentee is clicked on from the "your mentees" list from the home page, a mentee screen opens up showing that mentees information, tentatively pulling from a mentee array in "menteelist.js". Mood reports dynamically display based on "moodreports" value within each mentee item.